Fix PythonJob deserializer (#424)
- bump aiida-pythonjob to 0.1.8
- support register_pickle_by_value
- use the deserialization function from aiida-pythonjob
- override the `to_dict` method, to serialize the input data.
- move PythonJobTask to the task factory
