+++ pyproject.toml
@@ -33,7 +33,7 @@ dependencies = [
     "aiida-core>=2.3",
     "cloudpickle",
     "aiida-shell~=0.8",
-    "aiida-pythonjob>=0.1.6",
+    "aiida-pythonjob>=0.1.8",
     "fastapi",
     "uvicorn",
     "pydantic_settings",
